+++ src/samples/SimpleNotificationExample/GlobalUsings.cs
@@ -1,0 +1,9 @@
+global using System.Reflection;
+global using Microsoft.Extensions.DependencyInjection;
+global using Microsoft.Extensions.Hosting;
+global using Microsoft.Extensions.Logging;
+global using Blazing.Mediator;
+global using SimpleNotificationExample.Notifications;
+global using SimpleNotificationExample.Services;
+global using SimpleNotificationExample.Subscribers;
+global using SimpleNotificationExample.Middleware;